scripts/requirements.txt
pandas
python-docx
fpdf
faker
openpyxl
argparse
